requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
pymongo==3.11.4
